backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
passlib[bcrypt]
python-jose
pydantic
requests
